# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VentasFix API",
        "description": "API REST para gestión de Usuarios, Productos y Clientes de VentasFix",
        "version": "1.0.0"
    },
    "paths": {
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar todos los clientes",
                "operationId": "3ab1bd0bc1c6d13508917ec1f71f52f5",
                "responses": {
                    "200": {
                        "description": "Lista de clientes obtenida correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un nuevo cliente",
                "operationId": "22be7ccdd26efb899a619bd891090f4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rut_empresa",
                                    "rubro",
                                    "razon_social",
                                    "telefono",
                                    "direccion",
                                    "nombre_contacto",
                                    "email_contacto"
                                ],
                                "properties": {
                                    "rut_empresa": {
                                        "type": "string"
                                    },
                                    "rubro": {
                                        "type": "string"
                                    },
                                    "razon_social": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "nombre_contacto": {
                                        "type": "string"
                                    },
                                    "email_contacto": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener un cliente por su ID",
                "operationId": "e898290219910a22d5551f37ed2cfcaf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar un cliente por su ID",
                "operationId": "2dde92a886747eb383134d1ef3eedd33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rut_empresa": {
                                        "type": "string"
                                    },
                                    "rubro": {
                                        "type": "string"
                                    },
                                    "razon_social": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "nombre_contacto": {
                                        "type": "string"
                                    },
                                    "email_contacto": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado correctamente"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar un cliente por su ID",
                "operationId": "94088fd68be6ed57eec592470ce8ddbc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado correctamente"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum Token",
                "description": "Ingresa el token obtenido en /api/login, sin la palabra \"Bearer\" adelante."
            }
        }
    },
...
}